Revert "Merge branch 'main' into duyne"
This reverts commit f424741c323b74f468a6e2d5fb7ff418a23d0278, reversing
changes made to 9678225965bb2ee0613589606d7437d4fb4ea8ce.
