+++ .gitignore
@@ -1,3 +1,2 @@
 /build/
 /target/
-/src/main/java/utils/
