+++ src/main/java/DAO/AddressDAO.java
@@ -2,7 +2,6 @@ package DAO;
 
 import java.util.List;
 
-
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Join;
