+++ src/main/java/bean/Product.java
@@ -19,7 +19,9 @@ import javax.persistence.Table;
 @Table(name = "product")
 public class Product implements Serializable {
 
-	
+	/**
+	 * 
+	 */
 	private static final long serialVersionUID = 1L;
 
 	@Id
