+++ src/main/java/servlets/AboutServlet.java
@@ -1,4 +1,4 @@
-package servlets.shop;
+package servlets;
 
 import jakarta.servlet.RequestDispatcher;
 
@@ -9,16 +9,16 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet({"/shoping-cart"})
-public class ShopingCartServlet extends HttpServlet{
+@WebServlet({"/about"})
+public class AboutServlet extends HttpServlet{
 	private static final long serialVersionUID = 1L;
     
-    public ShopingCartServlet() {
+    public AboutServlet() {
         super();
     }
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		RequestDispatcher dispatcher = this.getServletContext().getRequestDispatcher("/views/shoping-cart.jsp");
+		RequestDispatcher dispatcher = this.getServletContext().getRequestDispatcher("/views/about.jsp");
 		
 		dispatcher.forward(request, response);
 	}
@@ -26,4 +26,5 @@ public class ShopingCartServlet extends HttpServlet{
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		doGet(request, response);
 	}
+
 }
