+++ src/main/java/servlets/ContactServlet.java
@@ -1,4 +1,4 @@
-package servlets.shop;
+package servlets;
 
 import jakarta.servlet.RequestDispatcher;
 
@@ -9,16 +9,16 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet({"/product-detail"})
-public class ProductDetailServlet extends HttpServlet{
+@WebServlet({"/contact"})
+public class ContactServlet extends HttpServlet{
 	private static final long serialVersionUID = 1L;
     
-    public ProductDetailServlet() {
+    public ContactServlet() {
         super();
     }
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		RequestDispatcher dispatcher = this.getServletContext().getRequestDispatcher("/views/product-detail.jsp");
+		RequestDispatcher dispatcher = this.getServletContext().getRequestDispatcher("/views/contact.jsp");
 		
 		dispatcher.forward(request, response);
 	}
@@ -26,4 +26,5 @@ public class ProductDetailServlet extends HttpServlet{
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		doGet(request, response);
 	}
+
 }
