+++ src/main/java/servlets/ForgotServlet.java
@@ -1,5 +1,6 @@
-package servlets.shop;
+package servlets;
 
+import jakarta.servlet.RequestDispatcher;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
@@ -8,19 +9,18 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet({"/logout"})
-public class LogoutServlet extends HttpServlet{
+@WebServlet({"/forgot"})
+public class ForgotServlet extends HttpServlet{
 	private static final long serialVersionUID = 1L;
     
-    public LogoutServlet() {
+    public ForgotServlet() {
         super();
     }
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		request.getSession().invalidate();
-
-		// Redrect to Home Page.
-		response.sendRedirect(request.getContextPath() + "/home");
+		RequestDispatcher dispatcher = this.getServletContext().getRequestDispatcher("/views/forgot.jsp");
+		
+		dispatcher.forward(request, response);
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
