+++ src/main/java/servlets/RegisterServlet.java
@@ -1,4 +1,4 @@
-package servlets.shop;
+package servlets;
 
 import java.io.IOException;
