+++ src/main/java/servlets/admin/PromotionsServlet.java
@@ -9,7 +9,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet({"/admin-promotions"})
+@WebServlet({"/promotions"})
 public class PromotionsServlet extends HttpServlet{
 	private static final long serialVersionUID = 1L;
