--- src/main/java/servlets/shop/HomeServlet.java
@@ -1,30 +1,0 @@
-package servlets.shop;
-
-import jakarta.servlet.RequestDispatcher;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@WebServlet({"/home"})
-public class HomeServlet extends HttpServlet {
-	private static final long serialVersionUID = 1L;
-       
-    public HomeServlet() {
-        super();
-    }
-
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		RequestDispatcher dispatcher = this.getServletContext().getRequestDispatcher("/views/index.jsp");
-		
-		dispatcher.forward(request, response);
-	}
-
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		doGet(request, response);
-	}
-
-}
