+++ src/main/webapp/views/admin/users.jsp
@@ -1,8 +1,5 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
-<%@page import="bean.UserAccount"%>
-
 <!DOCTYPE html>
 <html lang="vi">
 <head>
@@ -42,7 +39,7 @@
 				<hr class="sidebar-divider my-0" />
 				<ul class="navbar-nav text-light" id="accordionSidebar">
 					<li class="nav-item" style="margin-top: 35%"><a
-						class="nav-link" href="${pageContext.request.contextPath}/admin-dashboard"> <i
+						class="nav-link" href="admin-dashboard"> <i
 							class="fa fa-dashboard" style="font-size: 1.3rem"></i> <span
 							class="nav-item-content">Dashboard</span>
 					</a></li>
@@ -55,24 +52,24 @@
 							<div class="dropdown-menu">
 								<div class="dropdown-divider"></div>
 								<a class="dropdown-item text-white-50 menu-item"
-									href="${pageContext.request.contextPath}/admin-products">Products list</a> <a
+									href="admin-products">Products list</a> <a
 									class="dropdown-item text-white-50 menu-item"
-									href="${pageContext.request.contextPath}/admin-add-product">Add Product</a> <a
+									href="admin-add-product">Add Product</a> <a
 									class="dropdown-item text-white-50 menu-item"
-									href="${pageContext.request.contextPath}/admin-category">Category list</a>
+									href="admin-category">Category list</a>
 							</div>
 
 						</div>
 					</li>
-					<li class="nav-item"><a class="nav-link" href="${pageContext.request.contextPath}/admin-orders">
+					<li class="nav-item"><a class="nav-link" href="admin-orders">
 							<i class="fa fa-first-order" style="font-size: 1.3rem"></i> <span
 							class="nav-item-content">Orders</span>
 					</a></li>
 					<li class="nav-item"><a class="nav-link active"
-						href="${pageContext.request.contextPath}/admin-users"> <i class="fa fa-user"
+						href="admin-users"> <i class="fa fa-user"
 							style="font-size: 1.3rem"></i> <span class="nav-item-content">Users</span>
 					</a></li>
-					<li class="nav-item"><a class="nav-link" href="${pageContext.request.contextPath}/admin-promotions"><i
+					<li class="nav-item"><a class="nav-link" href="promotions"><i
 							class="fa fa-gift" style="font-size: 1.3rem"></i><span
 							class="nav-item-content">Promotions</span></a></li>
 				</ul>
@@ -107,17 +104,14 @@
 										class="d-flex d-xl-flex justify-content-xl-center align-items-xl-center dataTables_length my-2">
 									</div>
 									<a class="btn btn-info btn-sm w-25" role="button"
-										href="${pageContext.request.contextPath}/admin-add-user">ADD</a>
+										href="admin-add-user">ADD</a>
 								</div>
 							</div>
-							
-							<%-- List user --%>
 							<div class="table-responsive table mt-2" id="dataTable-1"
 								role="grid" aria-describedby="dataTable_info">
 								<table class="table my-0" id="dataTable">
 									<thead>
 										<tr>
-											<th class="text-uppercase"></th>
 											<th class="text-uppercase">name</th>
 											<th class="text-uppercase">email</th>
 											<th class="text-uppercase">Phone Number</th>
@@ -127,29 +121,38 @@
 										</tr>
 									</thead>
 									<tbody>
-										<c:forEach items="${UserList}" var="user" varStatus="i">
-											<tr>
-												<td>${i.index + 1}</td>
-												<td>${user.getName() }</td>
-												<td>${user.getEmailAddress() }</td>
-												<td>${user.getPhone_number() }</td>
-												<td>${user.getRole() }</td>
-												<td>
-													<div
-														class="border rounded-0 border-white d-flex justify-content-around btn-group">
-														<button class="btn" type="button" aria-expanded="false" data-toggle="dropdown"><i class="fa fa-ellipsis-v icon-size"></i></button>
-														<div class="dropdown-menu dropdown-menu-right" style="background: var(--white); position: static">
-															<a class="btn dropdown-item" role="button" href="${pageContext.request.contextPath}/admin-user-detail">Detail</a> 
-															<a class="btn dropdown-item" role="button" href="${pageContext.request.contextPath}/admin-delete-user">Delete</a>
-														</div>
+										<tr>
+											<td>Airi Satou</td>
+											<td>manager@gmail.com</td>
+											<td>0123456789</td>
+											<td>manager</td>
+											<td>
+												<div
+													class="border rounded-0 border-white d-flex justify-content-around btn-group">
+													<button class="btn" type="button" aria-expanded="false"
+														data-toggle="dropdown">
+														<i class="fa fa-ellipsis-v icon-size"></i>
+													</button>
+													<div class="dropdown-menu dropdown-menu-right"
+														style="background: var(--white); position: static">
+														<a class="btn dropdown-item" role="button"
+															href="admin-user-detail">Detail</a> <a
+															class="btn dropdown-item" role="button"
+															href="admin-delete-user">Delete</a>
 													</div>
-												</td>
-											</tr>
-										</c:forEach>
+												</div>
+											</td>
+										</tr>
 									</tbody>
 								</table>
 							</div>
-							
+							<div class="row">
+								<div
+									class="col-md-6 col-xl-12 d-xl-flex justify-content-xl-center">
+
+
+								</div>
+							</div>
 						</div>
 					</div>
 				</div>
